[JBRes-8782] some renaming
