+++ local-testing/tests/test_health.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from .utils import create_http_client
+from .idegym_utils import create_http_client
 
 
 @pytest.mark.asyncio
