+++ local-testing/tests/test_server_lifecycle.py
@@ -6,7 +6,7 @@ import pytest
 from idegym.api.orchestrator.servers import ServerReuseStrategy
 from kubernetes_asyncio.client import V1ResourceRequirements
 
-from .utils import create_http_client
+from .idegym_utils import create_http_client
 
 
 @pytest.mark.asyncio
