+++ local-testing/tests/test_server_strategies.py
@@ -7,7 +7,7 @@ from idegym.api.orchestrator.servers import ServerReuseStrategy
 from idegym.client.client import ServerCloseAction
 from kubernetes_asyncio.client import V1ResourceRequirements
 
-from .utils import create_http_client
+from .idegym_utils import create_http_client
 
 
 @pytest.mark.asyncio
